LGDP
Creation of the component "TermsConditionsModal", which has all the
information about the terms and conditions of the site.

Adding of the checkbox for the user to be able to accept the terms and
conditions in the pages "Cadastro" and "CadastroExterno".

Instalation of the library "react-cookies-consent", which was used to
build de consent message of the cookies in the page "Intranet".